# fix: resolve v9 and charting public docsite dependency issues and improve performance