--- a/apps/chart-docsite/.storybook/main.ts
+++ b/apps/chart-docsite/.storybook/main.ts
@@ -10,7 +10,7 @@ const config: StorybookConfig = {
     '../src/**/*.mdx',
     '../src/**/index.stories.@(js|jsx|ts|tsx)',
     // packages stories
-    '../../../packages/charts/*/stories/**/index.stories.@(js|jsx|ts|tsx)',
+    '../../../packages/charts/react-charts-preview/stories/**/index.stories.@(js|jsx|ts|tsx)',
   ],
 };