--- a/apps/chart-docsite/package.json
+++ b/apps/chart-docsite/package.json
@@ -2,7 +2,7 @@
   "name": "@fluentui/chart-docsite",
   "version": "1.0.0",
   "private": true,
-  "description": "Fluent UI React v9 documentation",
+  "description": "Fluent UI React Charts Preview documentation",
   "scripts": {
     "build-storybook": "storybook build -o ./dist/storybook --docs && node ./.storybook/fix-title.js 'Fluent UI Charts v9' ../dist",
     "clean": "just-scripts clean",
@@ -21,13 +21,10 @@
   },
   "dependencies": {
     "@fluentui/react-charts-preview": "*",
-    "@fluentui/react-components": "*",
-    "@griffel/react": "^1.5.22",
-    "@microsoft/applicationinsights-web": "^3",
+    "@fluentui/react-storybook-addon": "*",
+    "@fluentui/react-storybook-addon-export-to-sandbox": "*",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-window": "^1.8.6",
-    "tslib": "^2.1.0",
-    "react-hook-form": "^5.7.2"
+    "tslib": "^2.1.0"
   }
 }