--- a/apps/public-docsite-v9/src/DocsComponents/FluentDocsContainer.stories.tsx
+++ b/apps/public-docsite-v9/src/DocsComponents/FluentDocsContainer.stories.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
-import { DocsContainer, DocsContextProps } from '@storybook/addon-docs';
-import { FluentStoryContext } from '@fluentui/react-storybook-addon';
+import { DocsContainer, type DocsContextProps } from '@storybook/addon-docs';
+import { type FluentStoryContext } from '@fluentui/react-storybook-addon';
 import { webLightTheme, FluentProvider } from '@fluentui/react-components';
+
 
 interface FluentDocsContainerProps {
   context: FluentStoryContext & DocsContextProps;