--- a/apps/public-docsite-v9/src/DocsComponents/FluentDocsContainer.stories.tsx
+++ b/apps/public-docsite-v9/src/DocsComponents/FluentDocsContainer.stories.tsx
@@ -3,7 +3,6 @@ import { DocsContainer, type DocsContextProps } from '@storybook/addon-docs';
 import { type FluentStoryContext } from '@fluentui/react-storybook-addon';
 import { webLightTheme, FluentProvider } from '@fluentui/react-components';
 
-
 interface FluentDocsContainerProps {
   context: FluentStoryContext & DocsContextProps;
 }