--- a/change/@fluentui-chart-web-components-90a92ca1-c560-43e9-a5fa-cb87c11063ec.json
+++ b/change/@fluentui-chart-web-components-90a92ca1-c560-43e9-a5fa-cb87c11063ec.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: add new 'charting' domain tag",
+  "packageName": "@fluentui/chart-web-components",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "none"
+}