--- a/change/@fluentui-react-charting-a87fd601-35db-40d1-b7b6-be6b71ea83c2.json
+++ b/change/@fluentui-react-charting-a87fd601-35db-40d1-b7b6-be6b71ea83c2.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: add new 'charting' domain tag",
+  "packageName": "@fluentui/react-charting",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "none"
+}