--- a/change/@fluentui-react-charts-preview-cbb156e6-46e4-46d2-9772-6ba9c4778669.json
+++ b/change/@fluentui-react-charts-preview-cbb156e6-46e4-46d2-9772-6ba9c4778669.json
@@ -1,0 +1,7 @@
+{
+  "type": "none",
+  "comment": "chore: add new 'charting' domain tag",
+  "packageName": "@fluentui/react-charts-preview",
+  "email": "martinhochel@microsoft.com",
+  "dependentChangeType": "none"
+}