--- a/packages/charts/chart-web-components/project.json
+++ b/packages/charts/chart-web-components/project.json
@@ -3,7 +3,7 @@
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "implicitDependencies": [],
-  "tags": ["platform:web", "web-components"],
+  "tags": ["platform:web", "web-components", "charting"],
   "targets": {
     "e2e": { "dependsOn": ["build-storybook"] }
   }