--- a/packages/charts/react-charting/project.json
+++ b/packages/charts/react-charting/project.json
@@ -2,7 +2,7 @@
   "name": "react-charting",
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
-  "tags": ["v8", "ships-bundle"],
+  "tags": ["v8", "ships-bundle", "charting"],
   "implicitDependencies": [],
   "targets": {
     "test": {