--- a/packages/charts/react-charts-preview/library/project.json
+++ b/packages/charts/react-charts-preview/library/project.json
@@ -3,6 +3,6 @@
   "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "sourceRoot": "packages/charts/react-charts-preview/library/src",
-  "tags": ["platform:web", "vNext"],
+  "tags": ["platform:web", "vNext", "charting"],
   "implicitDependencies": []
 }