--- a/packages/react-components/react-storybook-addon/src/theme.ts
+++ b/packages/react-components/react-storybook-addon/src/theme.ts
@@ -1,5 +1,3 @@
-import type { Theme } from '@fluentui/react-theme';
-
 export const themes = [
   { id: 'web-light', label: 'Web Light' },
   { id: 'web-dark', label: 'Web Dark' },
@@ -12,5 +10,3 @@ export const defaultTheme = themes[0];
 
 export type ThemeIds = (typeof themes)[number]['id'];
 export type ThemeLabels = (typeof themes)[number]['label'];
-
-export type { Theme };