<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

**public-docsite-v9** 
- build target executes all react-component suite deps

**charting-docs**
- depends on whole react-components v9  
- depends on projectType: application, which is forbidden ( app cannot depend on another app )
- build target suffers same issue like docsite-v9

## New Behavior

**public-docsite-v9**

- build runs only necessary build targets of its dependencies 
  - from 69 to 14 target
  - speed bump 414s to 120s / **𝚫 71% FASTER**

**charting-docs**
- depends only on charting domain + shared workspace SB addons
- removes forbidden app dependencies
- build runs only necessary build targets of its dependencies 
  -  from 69 to 14 targets 
  -  speed bump from 361s to 75s / **𝚫 79% FASTER**

_BEFORE:_

<img width="2413" alt="image" src="https://github.com/user-attachments/assets/bbf76a79-074a-49ba-baec-7b3560d27302" />

<img width="848" alt="image" src="https://github.com/user-attachments/assets/74d3fcbf-9871-45e0-8869-412437f50720" />



_AFTER:_
<img width="1758" alt="image" src="https://github.com/user-attachments/assets/a23e4d12-4d20-4c43-a455-7d7f0ffb0ebd" />


**misc:**

- introduce new `charting` domain tag

## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Fixes #